Avoid std::string copy replace std::vector
The type of `p` on line 28 is `std::pair<const std::string, int>` so the call
to `move` on the next line will return `const std::string&` instead of an
r-value reference. This means that there is still an additional string copy
here. This commit stores pointers to the string to avoid this extra copy.

A new `struct` called `word_freq` has been created because it is a POD type
(unlike `std::pair<const std::string*, int>`). We replace the `std::vector`
with a dynamically allocated array to avoid calling the destructor of any
`word_freq` objects. Also creating the array with the correct size avoids
any reallocation and branching in `push_back`.