[clojure] add Makefile
I'm not quite happy with this, since it the time used to compile the
file into a java-class is counted in as time processing time in the
contest.

But I'll leave it as is for the following reasons:

* It was this way before considering the `run_commands.txt`
* I haven't found a way to do it better without introducing build tools

So the “maintainers” of clojure need to chime in here and find a proper
way of precompiling it. Or let everything as is, if it is the clojure
way.

Also I have removed downloading `clojure.jar` from `Dockerfile` and put
it in clojures `Makefile`. This way, it is not necessary to rebuilt the
image just to update it's version, also the clojure related JAR isn't
lingering in the project root anymore ;)